:construction: feat: api: 폴더 구조 재설정 - 엔드포인트별 폴더 생성
