+++ src/api/users/index.ts
@@ -1,0 +1,33 @@
+import { http } from "@/lib/http";
+import type * as T from "@/api/users/types";
+
+/* 회원가입 */
+export async function signUp(body: T.SignUpBody): Promise<T.User> {
+  const { data } = await http.post<T.User>("/users", body);
+  return data;
+}
+
+/* 로그인 */
+export async function login(body: T.LoginBody): Promise<T.User> {
+  const { data } = await http.post<T.User>("/users/login", body);
+  return data;
+}
+
+/* 사용자 정보 수정 */
+export async function updateUser(
+  userId: string,
+  body: T.UpdateUserBody
+): Promise<T.User> {
+  const { data } = await http.patch<T.User>(`/users/${userId}`, body);
+  return data;
+}
+
+/* 사용자 논리 삭제 */
+export async function deleteUser(userId: T.UserId): Promise<void> {
+  await http.delete<void>(`/users/${userId}`);
+}
+
+/* 사용자 물리 삭제 */
+export async function hardDeleteUser(userId: T.UserId): Promise<void> {
+  await http.delete<void>(`/users/${userId}/hard`);
+}
